requirement.txt
Django==5.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
django-filter==24.3
mysqlclient==2.2.4
Pillow==11.0.0
python-decouple==3.8